requirements.txt
fastmcp
pydantic
deepl
python-dotenv
airia
